Process - Add check on whether given or defined run_command exists.
Add safeguard to class and check whether the class has a defined
run_command as constant or is given as part of the initialization. A
defined constant will take precedence over a given command. If either of
the two are not found in the system PATH variable, a exception will be
thrown.