Stereo - Refactor accepted options for map projection.
Change the map_projected init argument and make the user specify the
type of session. When selecting one of the valid session types for map
projection, the alignment method option will be set to none. Passing in
this option will also trigger filtering of the run_options for those two
parameters.